feat: 74 - move link file into ui-config
