+++ src/components/registration/layout.ts
@@ -1,6 +1,6 @@
 import { REGISTRATION } from '../../pages/registration/constants';
-import { LINK_CONFIG } from '../../shared/components/link';
 import { HEADER2 } from '../../shared/styles';
+import { LINK_CONFIG } from '../../shared/ui-config/link';
 import { createA, createDiv, createH2 } from '../../utils/create-elements/create-tags';
 import { form } from './form';
