+++ src/shared/ui-config/link.ts
